Removes unneeded option, added err checking to port
Removes Unsecure as an option for meta. Unsecure is not needed as an
option for plugin authors since the meaning of it (from a snapd
perspective) has to do with legacy plugins and calling SetKey() or not.
SetKey() has no impact on grpc plugins so it makes sense to just
hardcode the value.

Adds error checking to port creation.